DEBUG-0.4.2: Changed loglevels: new loglevel IMPORTANT_ONLY replacing WARNING_AND_ERRORS_ONLY. Important log level includes scaling metrics causing scaling change

+++ src/utils/prometheusConnector.py
@@ -37,6 +37,6 @@ class PrometheusConnector:
         result = self._prometheusClient.custom_query(query=customized_memory_query)
 
         # Return memory value.
-        return int(result[0]['value'][1]) if result[0]['value'][1] else 0
+        return float(result[0]['value'][1]) if result[0]['value'][1] else 0
